Design : Modify css and move style components
