+++ pages/home/Home.tsx
@@ -12,13 +12,15 @@ import {
   hashtagContentsImages,
   popularContentsImages,
 } from '../../public/data';
-import { GetStaticProps, NextPage } from 'next';
+import { GetStaticProps, NextPage, GetServerSideProps } from 'next';
 import MobileBottomNav from '@/components/common/MobileBottomNav';
+import { useQuery } from 'react-query';
+import axios from 'axios';
 interface Props {
   data: SliderImage[];
   hashtagData: SliderImage[];
-  hashContent: SliderImage[];
-  popularContent: SliderImage[];
+  hashContent: any;
+  popularContent: any;
 }
 
 const Home: NextPage<Props> = ({
@@ -27,13 +29,14 @@ const Home: NextPage<Props> = ({
   hashContent,
   popularContent,
 }) => {
-  // console.log(data);
+  console.log(data);
+  console.log(popularContent);
   return (
     <Main>
       <Header />
-      <MobileBottomNav />
       <MainBannerSlider data={data} />
       <HashtagNavBar data={hashtagData} />
+      <MobileBottomNav />
       <HashtagContents data={hashContent} />
       <PopularContents data={popularContent} />
     </Main>
@@ -42,14 +45,28 @@ const Home: NextPage<Props> = ({
 
 export default Home;
 
+const MainPost = async () => {
+  const response = await axios.get(`http://localhost:3001/api/main`);
+  console.log(response);
+  return response.data;
+};
+MainPost();
+console.log(MainPost());
+
 export const getStaticProps: GetStaticProps = async () => {
+  const response = await axios.get(`http://localhost:3001/api/main`);
   const data = sliderImages;
   const hashtagData = hashtagImages;
-  const hashContent = hashtagContentsImages;
-  const popularContent = popularContentsImages;
+  // const hashContent = hashtagContentsImages;
+  // const popularContent = popularContentsImages;
   // console.log(data);
   return {
-    props: { data, hashtagData, hashContent, popularContent },
+    props: {
+      data,
+      hashtagData,
+      hashContent: response.data.posts,
+      popularContent: response.data.likePosts,
+    },
     revalidate: 60,
     //MEMO : revalidate
     //stale : 신선하지 않은 데이터
